Port XP curve from FIQHack
This is a mostly experimental change, but I believe that this XP curve
is better at allowing the character to keep gaining levels as the game
progresses past the midgame, rather than slaughtering everything in
Gehennom and /maybe/ reaching XL 15.

In this curve, levels 1-10 are the same as in vanilla, but after that
the required amount of XP is much less steep.