Add poltergeist monster.
This exists in the biodiversity patch, but I haven't looked at that
implementation. This implementation takes advantage of some 3.7
monmove changes and the addition of breaktest() for monster thrown
objects that I needed to make bottles work. I've tested this and
it't enormously fun / annoying, and feels very in the spirit of
NetHack. One automatically spawns in the valley of the dead because
why not welcome the player to Gehennom with something that makes
them tear their hair out?