Rename SoG to Scroll of Erasure.
You can view the entire discussion on NetHack's GitHub page, but the
long story short is that I'm one of the people of the opinion that
this name should have been changed a long time ago. While the new
name has its own problems, at the very least it's different, and
maybe it'll spark some discussion again. I'll be keeping the internal
references unchanged for now because I'd like to maintain 3.7 compatibility,
and running a simple find and replace is both impossible and NetHack
and likely to make this a nightmare to maintain.

At the end of the day, this is just a fork of an old game that
barely any of us play anymore. Please no one get made about this,
we all have way better things to be mad about.