Covetous monsters can assume normal movement when close to the player.
In the past, there's been several discussions concerning covetous
monster behavior, mostly centered around their movement. Which isn't
really moving in the normal sense, as they just teleport around all
over the place. This commit attempts to allow covetous monsters to
retain their covetous tactics and 'movement', but then being able to
move in a normal sense as well given the right conditions.

Per this graphic (the player is at the center as @):

.........
.........
..xxxxx..
..xxxxx..
..xx@xx..
..xxxxx..
..xxxxx..
.........
.........

The tile spots marked with an x is the distance from the player that a
covetous monster will bypass all of the warping behavior in tactics()
and use m_move() instead (normal movement). Per turn, there's a 5%
chance that a covetous monster using normal movement will revert back to
regular covetous warping.

Because of this, I'm now able to add the quest nemesis, demon lords and
demon princes to the 'monsters breaking boulders' routine. Boulder forts
are no longer a surefire method of completely protecting yourself from
various boss-type monsters. Boulder forts can still be effective
(there's a timeout for when breaking a boulder can happen again), but
it's no longer a simple affair.