New monster: Lightcrust.
This is something aosdict came up with that we spent some time
discussing. Born out of my desire to have more interactable light
sources in the dungeon. The radius is large, but given that they
are stationary and we are not in the 1990s this should not be an
issue.

May rename later?