Initial Pass: Harmonizing and Boosting.
Probably the largest gameplay change I will ever make, and what ties
the entire set of systems. The very, VERY short version of this change
is that you can get damage multipliers and AC boosts depending on the
terrain that you are on. This applies to monsters as well, however.

There are going to be many, many future alterations to this. It's
the main draw of Crecelle, and it's probably always going to need
balancing. For now, we want to get the code out there before all
else.