Fix: Dragons were dropping 0 weight scales. Fixes #29.
In mkobj.c:weight(), a small snippet was checking if for weapons and armor that didn't merge and returning obj->owt. However, the object wasn't fully initialized here so owt was always 0 for dragon scales (and potentially other objects).

After this fix, dragon scales were correctly weighing in at 40, however, another problem started where the fuzzing of weights wasn't occurring - at least for wished or items. I rearranged some of the weight fuzzing so that it should work both with random item generation and with wished for items.

A nice side note, player's starting inventory items seem to always be the same standard weight. This is probably good to discourage start scumming for optimal weights.