New Playable Race: Kobold
The appearance of kobolds as a playable race here is distinct
from their appearance in other variants. They are a challenge
race, in that their starting ability scores and score caps are
low, and they start with a kobold companion rather than a pet.

This commit also expands the potential playable races using some
code for mhflags from splice.