Switch from object harmonies system to oprops.
This is very similar to the oprops system across many variants, but
I rolled my own code. While there are not yet many oprops, it is
feature-complete, and includes specifying oprops in special level
lua, identifying oprops, and more. This commit also rips out the
remains of the harmony system and refactors boost flags as m4 flags.